+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.4
 * Improved Puppeteer/Chromium check for installed version and autoinstall
     * Added progress bar to download indicator
+* Update `markdown-elearnjs` to v1.6.1
 ## 1.0.3
 * Update `markdown-elearnjs` to v1.6.0
 ## 1.0.2